pyproject.toml
[tool.poetry]
name = "flox-sentry"
version = "0.1.4"
homepage = "https://github.com/getflox/flox-sentry"
description = "Automatically create projects and teams for flox managed projects"
authors = ["Michal Przytulski <michal@przytulski.pl>", "Daniel Ancuta <whisller@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
requests-toolbelt = "^0.10"
sentry-api-python = "^0.3.0"

[tool.poetry.plugins."flox.plugin"]
sentry = "flox_sentry:plugin"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
bump2version = "^1.0.1"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
